+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- The `interval` configuration option is now in milliseconds instead of seconds.
+
 ## [0.2.0] - 2026-09-06
 
 ### Added